Fixed a warning when compiling with -O3 on GCC 11.4.0
GCC previously complained that it was possible for an out-of-bounds
write to occur in the "buf" variable in "WriteULongest". Some
experimentation reveals that this happens when constant folding
a width of 0 or 1 into the function. Likely, the analysis fails
to realize that the loop will never run in these cases.